Handle secrets for HTTP tasks #CALM-13793
This handles basic auth and secret header variables for HTTP tasks.
We are using the existing basic auth for HTTP task, which is going to
change soon.
Keeping this in mind, we are using credential for basic auth in HTTP
task but injecting the secret in the auth object instead of just
adding a credential reference.

The following TODO has been added and should be addressed later:
Auth should be changed to basic auth with credential.
This is dependent on https://jira.nutanix.com/browse/CALM-12149
We could also possibly check calm server version to switch between
the two auth mechanisms since basic auth will be deprecated.

JIRA issue for this TODO item: https://jira.nutanix.com/browse/CALM-13856