Display message list on save of BP if in DRAFT state #CALM-14018
There is already some handling preset to display the messages on the
blueprint entity, but this does not include the messages on sub
entities. These changes pick up the messages from nested sub entities
and try to display it with a meaningful path to the entity that the
validation messages are present on.