Task/m calm 39476 add all projects update marketplace (#294)
https://jira.nutanix.com/browse/CALM-39476.
As of now users doesn't have an option to update the Marketplace runbook
or Marketplace blueprints with all_projects flag. As part of this review
made changes in order to support the update of marketplace runbook or
blueprints with all the projects in the system.

---------

Co-authored-by: Prabhat Dwivedi <prabhat.dwivedi@nutanix.com>
(cherry picked from commit 26255bc96ca0f133173cec154370a071104a16e3)