(#CALM-44134, #CALM-43590) - Fix VM Endpoint tests to pass account uuid and upload endpoint to default project (#415)
As part of the changes brought on by:
https://github.com/ideadevice/calm/pull/3269, a valid account reference
needs to be sent while creating VM endpoints. The test was sending just
a name and kind as part of the account_reference in the VM EP create
payload. Modified the test to send a valid account_reference and uuid in
the payload.

The tests also downloads and uploads the endpoint, but does it using the
_internal project. But since _internal project has no accounts in it.
the /endpoints/import_file was not seeding the account_uuid in the
uploaded endpoint. Hence modified the test to upload the endpoint to
default project instead of internal project.

---------

Co-authored-by: Prabhat Dwivedi <prabhat.dwivedi@nutanix.com>
(cherry picked from commit 18f153cbe0614b9a5fef68bb05c45bd23e24be4b)