CP Bug fixes - Part 1 (#464)
Bugs fixed as part of this PR:

1 - [CALM-45346](https://jira.nutanix.com/browse/CALM-45346)
In Decompile flows, regular CalmTask for type=HTTP doesn't support
inline URL fields. So, ensured that when Task template is rendered as
part of dynamic variables, CalmVarTask is used. Else the regular
CalmTask. Modified "user_attrs" & task jinja files as part of this
change

2 - [CALM-45348 ](https://jira.nutanix.com/browse/CALM-45348)
When basic auth WITHOUT creds are present in RTOpTask inargs' HTTP
variables, decompile is not creating a credential for the basic auth
details. Made changes to support this. As part of the same ticket, have
extended the same support to provider decompile flow as well.

3 - [CALM-45350](https://jira.nutanix.com/browse/CALM-45350)
When multiple RTs are present in the provider, a new line is not being
added between them during decompiled. Fixed this issue

4 - [CALM-45352](https://jira.nutanix.com/browse/CALM-45352)
As of now, there is no usecase for having "value" field for dynamic
variables in RB/BP as they are always set during Launch/Execute. But for
CP, values need to be provided for dynamic variables during RB save
itself (RTOpTask). Added changes to set the same for such runbooks
during decompile

5 - [CALM-45685](https://jira.nutanix.com/browse/CALM-45685)
This is an issue even with existing RB/BP flows. The JIRA description
mentions what the issue exactly is. To fix this, changed the request
body of HTTP tasks' decompile to use "repr(request_body)" instead of
json.dumps. Also, removed the default request body ("{}") being set for
DELETE tasks

(cherry picked from commit 8ba29a9ce2c3bde69dff2b3d29b463122f0eeb4d)