(#CALM-45118) runbook output variables as list of objects (#458)
As part of Custom Provider, we are adding support for output_variables
in provider actions. Changes for supporting output_variables as list of
objects instead of list of strings and supporting in provider actions as
well as user runbooks.

---------

Co-authored-by: Bekkam Venkata Aditya <bekkamvenkata.aditya@nutanix.com>
Co-authored-by: imranpasha1 <imran.pasha1@nutanix.com>
Co-authored-by: Utkarsh Bairolia <utkarsh.bairolia@nutanix.com>
Co-authored-by: Nirbhay Bagmar <nirbhay.bagmar@nutanix.com>
Co-authored-by: Richa Kumari <richa.kumari@nutanix.com>
Co-authored-by: Vartika Singh <vartika.singh>
(cherry picked from commit 9549604611cd177b3ed91dd1766ea9f068896aca)