(ENG-817125) Skip tunnel attribute if tunnel_reference is empty in blueprint create (#691)
**RCA:**
This PR addresses an issue where blueprint creation via DSL fails if the
`tunnel_reference` attribute is empty or None. The API expects
tunnel_reference to be either a valid object or omitted entirely, but
not None. Previously, an empty tunnel_reference was included in the
payload, resulting in a 422 error with the message: "None is not of type
'object'".

**Fix:**
The code now removes the `tunnel_reference` field from the payload if it
is empty or None, ensuring that only valid payload is sent in the API.

**Bug:** [ENG-817125](https://jira.nutanix.com/browse/ENG-817125)

**Test Plan:** Manually tested

(cherry picked from commit 61112725edae40fd112479bf5d3bd8a723cc1ed4)