(ENG-815337) Fix tunnel describe for tunnel with unicode characters in their names. (#713)
**__RCA:__**
The `get_tunnel()` function uses `get_name_query()` function which is
not supposed to as the backend does not support the filtering condition
formed by the `get_name_query` function.

**__Fix:__**
Added a simple filter for name string as we need to precisely search for
the name given in the command.
This is done at all the places in the repo, for reference I've attached
the places where it's currently being used instead of using the
`get_name_query` function:
1. Runbooks:
https://github.com/ideadevice/calm-dsl/blob/master/calm/dsl/cli/runbooks.py#L615
2. Endpoints:
https://github.com/ideadevice/calm-dsl/blob/master/calm/dsl/cli/endpoints.py#L188
3. Apps:
https://github.com/ideadevice/calm-dsl/blob/master/calm/dsl/cli/apps.py#L132
4. Blueprints:
https://github.com/ideadevice/calm-dsl/blob/master/calm/dsl/cli/bps.py#L841

(cherry picked from commit e0578384dcb42a9645fe173b2a3d68e7d7647ab7)