Fix for MPI BP launch is happening even if the global variable is not of MPI project (#738)
MPI BP launch is happening even if the global variable is not of MPI
project

Fix - In the backend marketplace_launch API, when a global variable
referenced in the MPI does not belong to the project used for the
launch, a validation error is added to the message_list. However, the
MPI BP state remains ACTIVE. As part of the fix, we now check the
message_list in the response for this specific error. If the error is
present, we stop the flow and raise an exception instead of allowing the
launch to proceed.

Ticket - [ENG-860396](https://jira.nutanix.com/browse/ENG-860396)

<img width="1134" height="193" alt="Screenshot 2025-11-26 at 3 55 43 PM"
src="https://github.com/user-attachments/assets/45355536-7e04-4e48-87c0-6ead19779f8e"
/>

(cherry picked from commit 89ee4705e1b31f2c7b756890f399b8a60f8088cc)